Collected endscreen functions in view/endscreen
